refactor(cde): add-property is a single inline form, not a nested wizard
Kill the wizard-inside-a-wizard. Extract the property builder into
PropertyForm.vue — one single-screen form with progressive disclosure
(Source toggle reveals CDE search+strength+options, or manual
type/values, or the bundle members table) and one "Add" action. No inner
step rail, no second counter.

- AddPropertyWizard becomes a thin el-dialog wrapper around PropertyForm
  (destroy-on-close → fresh each open); ModelSpecGenerator keeps using it
  unchanged.
- CreateModelWizard embeds PropertyForm inline in the Properties step
  (expands on "Add property", collapses on Add/Cancel), so the only step
  counter is the outer model wizard.
